+++ sdk/identity/Azure.Identity/src/ClientSecretCredential.cs
@@ -96,6 +96,7 @@ namespace Azure.Identity
                          tenantId,
                          clientId,
                          clientSecret,
+                         null,
                          options as ITokenCacheOptions,
                          (options as ClientSecretCredentialOptions)?.RegionalAuthority,
                          options?.IsLoggingPIIEnabled ?? false);